main/Faikin: Introduce uart_enabled()
A useful wrapper function instead of direct checking for tx.set && rx.set. On
ESP8266 pin assignment is static, so this function has different implementation
in that version.

proto_name() made to return "MOCK" if UART is disabled, so that dummy mode is
immediately made visible both in UI and debug log.

Signed-off-by: Pavel Fedin <pavel_fedin@mail.ru>