ESP: Introduce udp_discovery setting
It turns out that our legacy API isn't completely properly done; and it appears
to disrupt Daikin app's advanced functionality (like power saving modes),
causing it to crash.

Let's mitigate that by introducing a setting, which allows for enable or disable
auto-discovery on demand. Defaults to off since the feature is currently
experimental.

Signed-off-by: Pavel Fedin <pavel_fedin@mail.ru>